Set view tree owners in parent class
