+++ android/libraries/rib-android-core/src/main/kotlin/com/uber/rib/core/CoreAppCompatActivity.kt
@@ -20,6 +20,9 @@ import android.os.Bundle
 import androidx.annotation.CallSuper
 import androidx.annotation.IntRange
 import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.ViewTreeLifecycleOwner
+import androidx.lifecycle.ViewTreeViewModelStoreOwner
+import androidx.savedstate.ViewTreeSavedStateRegistryOwner
 
 /** Core Support v7 AppCompat Activity. */
 abstract class CoreAppCompatActivity : AppCompatActivity() {
@@ -32,6 +35,7 @@ abstract class CoreAppCompatActivity : AppCompatActivity() {
       activityDelegate = (applicationContext as HasActivityDelegate).activityDelegate()
     }
     super.onCreate(savedInstanceState)
+    initViewTreeOwners()
     activityDelegate?.onCreate(savedInstanceState)
   }
 
@@ -80,5 +84,17 @@ abstract class CoreAppCompatActivity : AppCompatActivity() {
   override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
     super.onActivityResult(requestCode, resultCode, data)
     activityDelegate?.onActivityResult(this, requestCode, resultCode, data)
+  }
+
+  /**
+   * [RibActivity] must call this since it does not use [ComponentActivity.setContentView] which
+   * already handles this.
+   */
+  private fun initViewTreeOwners() {
+    // Set the view tree owners before setting the content view so that the inflation process
+    // and attach listeners will see them already present
+    ViewTreeLifecycleOwner.set(window.decorView, this)
+    ViewTreeViewModelStoreOwner.set(window.decorView, this)
+    ViewTreeSavedStateRegistryOwner.set(window.decorView, this)
   }
 }
